Update README.md
Checking Intervals

Step 1

In college algebra we see numeric intervals given as

2 < x < 6

In C++ this open interval does not have the meaning you may expect.

Modify the code (fix the if statement) to give the correct C++ Boolean expression that specifies that the number lies between 2 and 6. We can think of this expression as:

number is greater than 2 AND number is less than 6

Step 2

We can also use Boolean expressions to see if a number falls outside of a range.

Add an additional if statement to specify if the input is less than 3 or greater than 7.